Moved the util methods into the test base class and reworked the expected change generation logic (#15)
